requirements.txt
selenium==4.8.0 
bs4==0.0.1 
webdriver-manager==3.8.5
psycopg2-binary==2.9.5 
pyyaml==6.0
schedule==1.1.0 